+++ src/components/button_component/ButtonComponent.h
@@ -40,7 +40,7 @@ namespace Project::Components {
     void build(Project::Utilities::LuaStateWrapper& luaStateWrapper, const std::string& tableName) override;
     void applyStyle() override;
 
-    void setEntityPosition(int x, int y) override;
+    void setEntityPosition(float x, float y) override;
     void setSize(int w, int h);
     
     void setColor(SDL_Color _color) { data.color = _color; }
